commit14

+++ src/views/commodity.py
@@ -4,7 +4,7 @@
 from flask import Blueprint, request, jsonify, send_file
 from uuid import uuid1
 import os
-from src.models import Commodity, db
+from src.models import Commodity, db, User
 from src.utils import img
 from src.security import token_auth
 
@@ -19,21 +19,21 @@ commodity_page = Blueprint('commodity_page', __name__)
 
 @commodity_page.route('/commodity', methods=['POST'])
 def add_commodity():
-    user_id = request.form['user_id']
-    area_id = int(request.form['area_id'])
-    batch = int(request.form['batch'])
-    name = request.form['name']
-    price = float(request.form['price'])
-    weight = float(request.form['weight'])
+    user_id = request.json['user_id']
+    area_id = int(request.json['area_id'])
+    batch = int(request.json['batch'])
+    name = request.json['name']
+    weight = float(request.json['weight'])
+    saler_id = float(request.json['saler_id'])
     logistics_id = str(uuid1())
-    ini = request.form['ini']
-    des = request.form['des']
+    ini = request.json['ini']
+    des = request.json['des']
 
-    qrcode_url = "http://127.0.0.1:5000" + "/purchase/" + str(logistics_id)
+    qrcode_url = "http://127.0.0.1:5000" + "/commodity/" + str(logistics_id) + '/' + str(saler_id)
     qr_path = os.path.dirname(os.path.dirname(os.path.dirname(__file__))) + '/static/qr_codes/'+logistics_id+'.png'
     img.make_qrcode(qrcode_url, qr_path)
 
-    commodity = Commodity(user_id, area_id, batch, name, price, weight, logistics_id, ini, des, qrcode_url)
+    commodity = Commodity(user_id, area_id, batch, name, weight, saler_id, logistics_id, ini, des, qrcode_url)
     db.session.add(commodity)
     db.session.commit()
 
@@ -47,13 +47,20 @@ def add_commodity():
 
 @commodity_page.route('/commodity/<logistics_id>', methods=['DELETE'])
 def del_commodity(logistics_id):
-    commodity = Commodity.query.filter(Commodity.logistics_id == logistics_id).first()
-    if commodity:
-        db.session.delete(commodity)
-        db.session.commit()
-        return '删除成功'
+    token_data = token_auth.verify_token(request.headers['token'])
+    if token_data == 'token过期或错误':
+        return '请重新登录'
+    role = User.query.filter(User.user_id == token_data['user_id']).first().role
+    if role == 'producer' or 'admin':
+        commodity = Commodity.query.filter(Commodity.logistics_id == logistics_id).first()
+        if commodity:
+            db.session.delete(commodity)
+            db.session.commit()
+            return '删除成功'
+        else:
+            return '商品不存在'
     else:
-        return '商品不存在'
+        return '无权限'
 
 
 '''
@@ -63,22 +70,28 @@ def del_commodity(logistics_id):
 
 @commodity_page.route('/commodity', methods=['GET'])
 def query_commodity():
-    commodities = Commodity.query.all()
-    data = []
-    for commodity in commodities:
-        data.append({
-            'user_id': commodity.user_id,
-            'area_id': commodity.area_id,
-            'batch': commodity.batch,
-            'name': commodity.name,
-            'price': commodity.price,
-            'weight': commodity.weight,
-            'logistics_id': commodity.logistics_id,
-            'ini': commodity.ini,
-            'des': commodity.des,
-            'qrcode_url': commodity.qrcode_url
-        })
-    return jsonify(data)
+    token_data = token_auth.verify_token(request.headers['token'])
+    if token_data == 'token过期或错误':
+        return '请重新登录'
+    if token_data['user_id'] == '0':
+        commodities = Commodity.query.all()
+        data = []
+        for commodity in commodities:
+            data.append({
+                'user_id': commodity.user_id,
+                'area_id': commodity.area_id,
+                'batch': commodity.batch,
+                'name': commodity.name,
+                'price': commodity.price,
+                'weight': commodity.weight,
+                'logistics_id': commodity.logistics_id,
+                'ini': commodity.ini,
+                'des': commodity.des,
+                'qrcode_url': commodity.qrcode_url
+            })
+        return jsonify(data)
+    else:
+        return '无权限'
 
 
 '''
@@ -88,22 +101,28 @@ def query_commodity():
 
 @commodity_page.route('/commodity/<logistics_id>', methods=['GET'])
 def get_commodity(logistics_id):
-    commodity = Commodity.query.filter(Commodity.logistics_id == logistics_id).first()
-    if commodity:
-        return jsonify({
-            'user_id': commodity.user_id,
-            'area_id': commodity.area_id,
-            'batch': commodity.batch,
-            'name': commodity.name,
-            'price': commodity.price,
-            'weight': commodity.weight,
-            'logistics_id': commodity.logistics_id,
-            'ini': commodity.ini,
-            'des': commodity.des,
-            'qrcode_url': commodity.qrcode_url
-        })
+    token_data = token_auth.verify_token(request.headers['token'])
+    if token_data == 'token过期或错误':
+        return '请重新登录'
+    if token_data['user_id'] == '0':
+        commodity = Commodity.query.filter(Commodity.logistics_id == logistics_id).first()
+        if commodity:
+            return jsonify({
+                'user_id': commodity.user_id,
+                'area_id': commodity.area_id,
+                'batch': commodity.batch,
+                'name': commodity.name,
+                'price': commodity.price,
+                'weight': commodity.weight,
+                'logistics_id': commodity.logistics_id,
+                'ini': commodity.ini,
+                'des': commodity.des,
+                'qrcode_url': commodity.qrcode_url
+            })
+        else:
+            return '商品不存在'
     else:
-        return '商品不存在'
+        return '无权限'
 
 
 '''
@@ -111,8 +130,8 @@ def get_commodity(logistics_id):
 '''
 
 
-@commodity_page.route('/commodity/img/<logistics_id>', methods=['GET'])
-def get_img(logistics_id):
+@commodity_page.route('/commodity/qrcode_img/<logistics_id>', methods=['GET'])
+def get_qrcode(logistics_id):
     img_path = os.path.dirname(os.path.dirname(os.path.dirname(__file__))) + '/static/qr_codes/'+logistics_id+'.png'
     return send_file(img_path, mimetype='image/gif')
